pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "itsm-openenv-benchmark"
version = "0.1.0"
description = "Deterministic ITSM OpenEnv benchmark with 181 tasks"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "openenv-core>=0.2.0",
  "fastapi>=0.111.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.7.0",
  "requests>=2.31.0",
  "openai>=1.35.0",
  "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "ipykernel>=7.2.0",
  "matplotlib>=3.9.0",
  "pytest>=8.2.0",
  "ruff>=0.6.0",
]

[project.scripts]
server = "server.app:main"
itsm-openenv-server = "server.app:main"
itsm-openenv-infer = "inference:main"

[tool.setuptools.packages.find]
include = ["itsm_openenv_benchmark*", "server*"]
namespaces = false
